+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changes to be released in next version
 
 🙌 Improvements
  * RoomMemberDetailsVC: Enable / disable "Hide all messages from this user" from settings (#4281).
+ * SettingsVC: Show / hide NSFW and decrypted content options from build settings (#4290).
 
 🐛 Bugfix
  * RoomVC: Avoid navigation to integration management using integration popup with settings set to integration disabled (#4261).